feat: remove platform config generation in "kit apply"
This will be replaced with a templating mechanism from a bootstrap module.
Refactor to share immplementation with "kit bundle".
